requirements.txt
streamlit>=1.25.0
openai==0.27.0
pinecone-client>=2.2.2
sentence-transformers>=2.2.2
gitpython>=3.1.31
langchain>=0.0.208
python-dotenv>=1.0.0
